+++ src/test/java/carracing/view/StringUtilsTest.java
@@ -1,4 +1,4 @@
-package carracing.util;
+package carracing.view;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
